INTYGFV-11640: Dölj valideringen om man tömmer fältet och visa först när man har skrivit i något
